More compatibility changes for 64 bit, and cleaning up a few things.
git-svn-id: https://svn.code.sf.net/p/personalrobots/code/pkg/trunk@18781 f5854215-dd47-0410-b2c4-cdd35faa7885

Former-commit-id: 52668d46e84dd4337c1ea005eda869d1db2f884d

+++ sandbox/descriptors/descriptors.h
@@ -200,7 +200,7 @@ class SuperpixelStatistic : public ImageDescriptor {
   SuperpixelStatistic(int seed_spacing, float scale, SuperpixelStatistic* provider);
   //! Computes superpixels and puts into seg_, and computes the superpixel to pixel index.  Is called automatically, if necessary, by the compute(.) function.
   void segment();
-  IplImage* createSegmentMask(long label, CvRect* rect);
+  IplImage* createSegmentMask(int label, CvRect* rect);
 };
 
 class SuperpixelColorHistogram : public SuperpixelStatistic {
@@ -228,8 +228,8 @@ class SuperpixelColorHistogram : public SuperpixelStatistic {
   void clearPointCache() {}
   void clearImageCache();
   ~SuperpixelColorHistogram();
-  void computeHistogram(long label);
-  void computeHistogramCV(long label); 
+  void computeHistogram(int label);
+  void computeHistogramCV(int label); 
 };
 
 std::vector<ImageDescriptor*> setupImageDescriptors();
